+++ src/main/java/calculator/exception/calculator/CalculatorException.java
@@ -1,4 +1,4 @@
-package calculator.exception.input;
+package calculator.exception.calculator;
 
 public class CalculatorException extends IllegalArgumentException {
